Project Files
Endless Forest or infinite looping game is a subgenre of platform game in which the player character for an infinite amount of time while avoiding obstacles. The player's objective is to reach a high score by surviving for as long as possible. Its popularity is attributed to its simple gameplay that works well on touchscreen devices.